refactor: Project Architecture to Domain Driven Design, User Domain classes and authentication
