+++ src/main/java/br/com/fiap/irrigationapi/config/RequestFilter.java
@@ -1,0 +1,41 @@
+package br.com.fiap.irrigationapi.config;
+
+import br.com.fiap.irrigationapi.modules.users.repositories.UserRepository;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+import java.io.IOException;
+
+
+@Component
+public class RequestFilter extends OncePerRequestFilter {
+
+    @Autowired
+    TokenService tokenService;
+
+    @Autowired
+    UserRepository repository;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+        var authorizationHeader = request.getHeader("Authorization");
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            var token = authorizationHeader.replace("Bearer ", "").trim();
+            var login = tokenService.validateToken(token);
+            var user = repository.findByEmail(login);
+            var authenticationToken = new UsernamePasswordAuthenticationToken (user, null, user.getAuthorities());
+
+            SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+        }
+
+        filterChain.doFilter(request,response);
+    }
+}
