+++ src/main/java/br/com/fiap/irrigationapi/config/SecurityConfig.java
@@ -1,0 +1,83 @@
+package br.com.fiap.irrigationapi.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+
+
+    @Autowired
+    private RequestFilter requestFilter;
+
+    @Bean
+    public SecurityFilterChain filterChain(
+            HttpSecurity httpSecurity) throws Exception {
+
+        return httpSecurity.csrf(csrf -> csrf.disable())
+                .sessionManagement(
+                        session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+                .authorizeHttpRequests(authorize -> authorize
+
+                        // AREAS
+                        // TO DO: IMPLEMENTS
+
+                        // NOTIFICATIONS
+                        // TO DO: IMPLEMENTS
+
+                        // SCHEDULES
+                        // TO DO: IMPLEMENTS
+
+                        // SENSORS
+                        // TO DO: IMPLEMENTS
+
+                        // USERS
+                        .requestMatchers(HttpMethod.POST, "/api/users/register").permitAll()
+                        .requestMatchers(HttpMethod.POST, "/api/users/login").permitAll()
+                        .requestMatchers(HttpMethod.POST, "/api/users/login").permitAll()
+                        .requestMatchers(HttpMethod.POST, "/api/users/update-info").hasAnyRole("ADMIN", "USER")
+                        .requestMatchers(HttpMethod.POST, "/api/users/activate").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.POST, "/api/users/deactivate").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.POST, "/api/users/upgrade-permission").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.POST, "/api/users/downgrade-permission").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.GET, "/api/users/id").hasAnyRole("ADMIN", "USER")
+                        .requestMatchers(HttpMethod.GET, "/api/users/email").hasAnyRole("ADMIN", "USER")
+                        .requestMatchers(HttpMethod.GET, "/api/users/all").hasRole("ADMIN")
+
+                        // WEATHERS
+                        // TO DO: IMPLEMENTS
+
+                        .anyRequest().authenticated()
+                )
+                .addFilterBefore(
+                        requestFilter,
+                        UsernamePasswordAuthenticationFilter.class
+                )
+                .build();
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(
+            AuthenticationConfiguration authenticationConfiguration) throws Exception {
+        return authenticationConfiguration.getAuthenticationManager();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
+
