--- src/main/java/br/com/fiap/irrigationapi/config/secutity/VerificarToken.java
@@ -1,42 +1,0 @@
-package br.com.fiap.irrigationapi.config.secutity;
-
-import br.com.fiap.irrigationapi.repository.UserRepository;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class VerificarToken extends OncePerRequestFilter {
-
-    @Autowired
-    TokenService tokenService;
-
-    @Autowired
-    UserRepository userRepository;
-
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-
-        var authorizationHeader = request.getHeader("Authorization");
-
-        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
-            var token = authorizationHeader.replace("Bearer ", "").trim();
-            var login = tokenService.validateToken(token);
-            var user = userRepository.findByEmail(login);
-            var authenticationToken = new UsernamePasswordAuthenticationToken (user, null, user.getAuthorities());
-
-            SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-        }
-
-        filterChain.doFilter(request,response);
-    }
-}
