--- src/main/java/br/com/fiap/irrigationapi/controller/AreaController.java
@@ -1,18 +1,0 @@
-//package br.com.fiap.irrigationapi.controller;
-//
-//import br.com.fiap.irrigationapi.service.AreaService;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RestController;
-//
-//@RestController
-//@RequestMapping("/api/areas")
-//public class AreaController {
-//
-//    @Autowired
-//    private AreaService service;
-//
-//
-//    // TO DO: Implements routes
-//
-//}
