--- src/main/java/br/com/fiap/irrigationapi/controller/NotificationController.java
@@ -1,17 +1,0 @@
-package br.com.fiap.irrigationapi.controller;
-
-import br.com.fiap.irrigationapi.service.NotificationService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/notifications")
-public class NotificationController {
-
-    @Autowired
-    private NotificationService service;
-
-    // TO DO: Implements routes
-
-}
