--- src/main/java/br/com/fiap/irrigationapi/dto/LoginOutput.java
@@ -1,4 +1,0 @@
-package br.com.fiap.irrigationapi.dto;
-
-public record LoginOutput(String token) {
-}
