--- src/main/java/br/com/fiap/irrigationapi/model/Notification.java
@@ -1,42 +1,0 @@
-package br.com.fiap.irrigationapi.model;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import java.time.LocalDateTime;
-import java.util.Objects;
-
-@Entity
-@Table(name = "tbl_notification")
-@Getter
-@Setter
-public class Notification {
-
-    @Id
-    @Column(name = "notification_id")
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_NOTIFICATION")
-    @SequenceGenerator(name = "SEQ_NOTIFICATION", sequenceName = "SEQ_NOTIFICATION", allocationSize = 1)
-    private Long id;
-
-    private String description;
-
-    private LocalDateTime timestamp;
-
-    @ManyToOne
-    @JoinColumn(name = "sensor_id")
-    private Sensor sensor;
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Notification that = (Notification) o;
-        return Objects.equals(id, that.id) && Objects.equals(description, that.description) && Objects.equals(timestamp, that.timestamp) && Objects.equals(sensor, that.sensor);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, description, timestamp, sensor);
-    }
-}
