--- src/main/java/br/com/fiap/irrigationapi/model/UserRole.java
@@ -1,14 +1,0 @@
-package br.com.fiap.irrigationapi.model;
-
-public enum UserRole {
-    ADMIN( "admin"),
-    USER ("user");
-
-    private final String _role;
-    public String getRole() { return _role; }
-
-    UserRole (String role) {
-        this._role = role;
-    }
-
-}
