+++ src/main/java/br/com/fiap/irrigationapi/modules/areas/repositories/AreaRepository.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.areas.repositories;
+
+import br.com.fiap.irrigationapi.modules.areas.models.Area;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AreaRepository extends JpaRepository<Area, Long> {
+
+    // TO DO: Implements
+
+}
