+++ src/main/java/br/com/fiap/irrigationapi/modules/areas/services/AreaService.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.areas.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AreaService {
+
+    // TO DO: Implements
+
+}
