+++ src/main/java/br/com/fiap/irrigationapi/modules/notifications/models/Notification.java
@@ -1,0 +1,31 @@
+package br.com.fiap.irrigationapi.modules.notifications.models;
+
+import br.com.fiap.irrigationapi.modules.sensors.models.Sensor;
+import jakarta.persistence.*;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "tbl_notification")
+@Getter
+@Setter
+@EqualsAndHashCode
+public class Notification {
+
+    @Id
+    @Column(name = "notification_id")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_NOTIFICATION")
+    @SequenceGenerator(name = "SEQ_NOTIFICATION", sequenceName = "SEQ_NOTIFICATION", allocationSize = 1)
+    private Long id;
+
+    private String description;
+
+    private LocalDateTime timestamp;
+
+    @ManyToOne
+    @JoinColumn(name = "sensor_id")
+    private Sensor sensor;
+
+}
