+++ src/main/java/br/com/fiap/irrigationapi/modules/notifications/repositories/NotificationRepository.java
@@ -1,0 +1,9 @@
+package br.com.fiap.irrigationapi.modules.notifications.repositories;
+
+import br.com.fiap.irrigationapi.modules.notifications.models.Notification;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
+
+    // TO DO: Implements
+}
