+++ src/main/java/br/com/fiap/irrigationapi/modules/notifications/services/NotificationService.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.notifications.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class NotificationService {
+
+    // TO DO: Implements
+
+}
