+++ src/main/java/br/com/fiap/irrigationapi/modules/sensors/repositories/SensorRepository.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.sensors.repositories;
+
+import br.com.fiap.irrigationapi.modules.sensors.models.Sensor;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SensorRepository extends JpaRepository<Sensor, Long> {
+
+    // TO DO: Implements
+
+}
