+++ src/main/java/br/com/fiap/irrigationapi/modules/sensors/services/SensorService.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.sensors.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SensorService {
+
+    // TO DO: Implements
+
+}
