+++ src/main/java/br/com/fiap/irrigationapi/modules/users/dtos/LoginOutput.java
@@ -1,0 +1,4 @@
+package br.com.fiap.irrigationapi.modules.users.dtos;
+
+public record LoginOutput(String token) {
+}
