+++ src/main/java/br/com/fiap/irrigationapi/modules/users/dtos/UserOutput.java
@@ -1,0 +1,20 @@
+package br.com.fiap.irrigationapi.modules.users.dtos;
+
+import br.com.fiap.irrigationapi.modules.users.models.User;
+import br.com.fiap.irrigationapi.modules.users.models.UserRole;
+
+public record UserOutput(
+    Long id,
+    String name,
+    String email,
+    UserRole role
+){
+    public UserOutput(User user) {
+        this(
+                user.getId(),
+                user.getName(),
+                user.getEmail(),
+                user.getRole()
+        );
+    }
+}
