+++ src/main/java/br/com/fiap/irrigationapi/modules/users/models/UserRole.java
@@ -1,0 +1,14 @@
+package br.com.fiap.irrigationapi.modules.users.models;
+
+public enum UserRole {
+    ADMIN( "admin"),
+    USER ("user");
+
+    private final String _role;
+    public String getRole() { return _role; }
+
+    UserRole (String role) {
+        this._role = role;
+    }
+
+}
