+++ src/main/java/br/com/fiap/irrigationapi/modules/weathers/services/WeatherService.java
@@ -1,0 +1,10 @@
+package br.com.fiap.irrigationapi.modules.weathers.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class WeatherService {
+
+    // TO DO: Implements
+
+}
