--- src/main/java/br/com/fiap/irrigationapi/repository/AreaRepository.java
@@ -1,10 +1,0 @@
-//package br.com.fiap.irrigationapi.repository;
-//
-//import br.com.fiap.irrigationapi.model.Area;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface AreaRepository extends JpaRepository<Area, Long> {
-//
-//    // TO DO: Implements
-//
-//}
