--- src/main/java/br/com/fiap/irrigationapi/repository/NotificationRepository.java
@@ -1,9 +1,0 @@
-package br.com.fiap.irrigationapi.repository;
-
-import br.com.fiap.irrigationapi.model.Notification;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
-
-    // TO DO: Implements
-}
