--- src/main/java/br/com/fiap/irrigationapi/repository/SensorRepository.java
@@ -1,10 +1,0 @@
-package br.com.fiap.irrigationapi.repository;
-
-import br.com.fiap.irrigationapi.model.Sensor;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SensorRepository extends JpaRepository<Sensor, Long> {
-
-    // TO DO: Implements
-
-}
