--- src/main/java/br/com/fiap/irrigationapi/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package br.com.fiap.irrigationapi.repository;
-
-import br.com.fiap.irrigationapi.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    UserDetails findByEmail(String email);
-
-}
