--- src/main/java/br/com/fiap/irrigationapi/service/AreaService.java
@@ -1,10 +1,0 @@
-//package br.com.fiap.irrigationapi.service;
-//
-//import org.springframework.stereotype.Service;
-//
-//@Service
-//public class AreaService {
-//
-//    // TO DO: Implements
-//
-//}
