--- src/main/java/br/com/fiap/irrigationapi/service/SensorService.java
@@ -1,10 +1,0 @@
-package br.com.fiap.irrigationapi.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class SensorService {
-
-    // TO DO: Implements
-
-}
