--- src/main/java/br/com/fiap/irrigationapi/service/WeatherService.java
@@ -1,10 +1,0 @@
-package br.com.fiap.irrigationapi.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class WeatherService {
-
-    // TO DO: Implements
-
-}
